Grapple snapback
plunger leaves the gun on activation and snap back to barrel with deactivation
