+++ Assets/Scripts/Arrow Scripts/Phase1Trigger.cs
