+++ Assets/Scripts/Arrow Scripts/Phase2Trigger.cs.meta
