+++ Assets/Scripts/Lock&Key/LockTriggerCondition.cs.meta
